Move HTTP specifics to HttpResourceDataElement from node
